+++ libs/ledger-state/src/Cardano/Ledger/State/Vector.hs
@@ -10,14 +10,14 @@ module Cardano.Ledger.State.Vector where
 import Cardano.Ledger.Coin
 import Cardano.Ledger.Credential
 import Cardano.Ledger.Keys as Keys
-import Cardano.Ledger.State (StakePoolParams)
+import Cardano.Ledger.State (StakePoolSnapShot)
 import Control.DeepSeq
 import Data.Map.Strict as Map
 
 data SnapShotM = SnapShotM
   { ssStake :: !(Map (Credential Staking) (CompactForm Coin))
   , ssDelegations :: !(Map (Credential Staking) (KeyHash StakePool))
-  , ssStakePoolParams :: !(Map (KeyHash StakePool) StakePoolParams)
+  , ssStakePoolsSnapShot :: !(Map (KeyHash StakePool) StakePoolSnapShot)
   }
 
 instance NFData SnapShotM where